weechat-rs: Turn some common methods to static ones.
Some Weechat methods return values that don't need to be bound to a
lifetime of a Weechat context, those can be safely called anytime while
inside the Weechat thread.

We check at call time the thread ID to make sure that they are called in
the main Weechat thread since they are not thread safe.

This makes it possible to call Weechat::print() safely and without any
hacks from a Future that is running on the main Weechat thread using
the Weechat::spawn() method.