weechat: Don't drop the plugin pointer and thread id.
When a plugin gets unloaded buffer close callbacks of the plugin will
get called after the unload callback is called.

This means that some buffer close callbacks might want to print out
something or do some other weechat calls after the unload callback is
done.

Unconditionally setting the pointer and thread id in the init method and
never droping them seems to be the right approach.