buffer: Move the nick removing method under the buffer.
The buffer owns nicks, since you can take out multiple handles to the
same Nick object using the search_nick() method, removing a nick while
consuming the nick object won't guarantee that all handles are
invalidated.

Nick object holds a reference to a buffer using phantom data, meaning if
we move the nick removing method under the buffer and mark it as taking
a mutable reference to the buffer holding on to any Nick objects while
removing a nick will be disallowed.